refactor: extract post header to common
